CSharpDocumentPropertiesService Refactor
